+++ tests/test_countries.py
@@ -63,11 +63,14 @@ class TestCountries:
             ("Heard Island", {}, "HM"),  # en
             ("Norfolk Island", {}, "NF"),  # en
             ("Solomon Islands", {}, "SB"),  # en
-            ("Prince Edward Island", {}, "CA"),  # en
-            ("Rhode Island", {}, "US"),  # en
             # However, in cases where island is singular instead of plural,
             # there can be confusion.
-            ("Solomon Island Nationals", {}, "IS"),  # en
+            ("Solomon Island Nationals", {}, None),  # en
+            ("Prince Edward Island", {}, "CA"),  # en
+            ("Rhode Island", {}, "US"),  # en
+            ("Pacific island", {}, None),  # en
+            ("Fiji/Pacific island", {}, "FJ"),  # en
+            ("island", {}, None),  # en
             # Any capitalization for lang works
             ("Germany", {"lang": "en"}, "DE"),
             ("Germany", {"lang": "En"}, "DE"),
@@ -164,6 +167,7 @@ class TestCountries:
             ("state of palestine", "PS"),
             ("palestine, state of", "PS"),
             ("Irak", None),
+            ("island", None),
         ],
     )
     def test_country_name_to_country_code_en(self, input_data, expected):
@@ -191,6 +195,8 @@ class TestCountries:
             ("   Land der Hinrichtung : Deutschland", "DE"),
             ("Dschibuti-Stadt", "DJ"),
             ("elfenbeinküste und ecuador    ", "CI"),
+            ("Fiji/Pacific island", None),  # en
+            ("island", "IS"),
         ],
     )
     def test_country_name_to_country_code_de(self, input_data, expected):